// src/lib/game/modes/avalon/actions/types.ts
export enum AvalonAction {
  AdvanceFromQuest = "advance-from-quest",
  AdvanceFromTeamVote = "advance-from-team-vote",
  CastTeamVote = "cast-team-vote",
  PlayQuestCard = "play-quest-card",
  ProposeTeam = "propose-team",
  ResolveAssassination = "resolve-assassination",
  ResolveQuest = "resolve-quest",
  ResolveTeamVote = "resolve-team-vote",
  SelectAssassinationTarget = "select-assassination-target",
}